Abstract for the views
