+++ edu.rice.cs.hpcviewer.ui/Application.e4xmi
@@ -125,7 +125,7 @@
   <descriptors xmi:id="_4LP-wK0hEeqX1ux1mZi76A" elementId="edu.rice.cs.hpcviewer.ui.partdescriptor.graph.sort" label="Sorted plot graph" iconURI="platform:/plugin/edu.rice.cs.hpcviewer.ui/resources/hpcviewer-icon-editor-plotsort.png" bindingContexts="_c3AF2cjYEeSyMNYR5xypkQ" allowMultiple="true" category="editor" closeable="true" contributionURI="bundleclass://edu.rice.cs.hpcviewer.ui/edu.rice.cs.hpcviewer.ui.graph.GraphPlotSortViewer">
     <tags>categoryTag:hpcview</tags>
   </descriptors>
-  <descriptors xmi:id="_uSUdoMizEeq3_poNktUMtQ" elementId="edu.rice.cs.hpcviewer.ui.partdescriptor.profile" label="Profile" iconURI="platform:/plugin/edu.rice.cs.hpcviewer.ui/resources/hpcviewerEclipse32.gif" tooltip="Profile view" allowMultiple="true" contributionURI="bundleclass://edu.rice.cs.hpcviewer.ui/edu.rice.cs.hpcviewer.ui.parts.ProfilePart">
+  <descriptors xmi:id="_uSUdoMizEeq3_poNktUMtQ" elementId="edu.rice.cs.hpcviewer.ui.partdescriptor.profile" label="Profile view" iconURI="platform:/plugin/edu.rice.cs.hpcviewer.ui/resources/hpcviewerEclipse16.gif" tooltip="Profile view" allowMultiple="true" contributionURI="bundleclass://edu.rice.cs.hpcviewer.ui/edu.rice.cs.hpcviewer.ui.parts.ProfilePart">
     <toolbar xmi:id="_1e6AIMizEeq3_poNktUMtQ" elementId="edu.rice.cs.hpcviewer.ui.toolbar.0"/>
   </descriptors>
   <descriptors xmi:id="_0egRoMk4Eeq3_poNktUMtQ" elementId="edu.rice.cs.hpcviewer.ui.partdescriptor.trace" label="Trace" iconURI="platform:/plugin/edu.rice.cs.hpcviewer.ui/resources/hpc-merged-16x16.png" allowMultiple="true" contributionURI="bundleclass://edu.rice.cs.hpcviewer.ui/edu.rice.cs.hpcviewer.ui.parts.TracePart">
